+++ src/serena/utilities/local_minima_variation.py
@@ -4,18 +4,25 @@ File to hold the local minima variation code
 
 from datetime import datetime
 from typing import List, Dict
+from dataclasses import dataclass
 
-from serena.utilities.ensemble_groups import MultipleEnsembleGroups
+from serena.utilities.ensemble_groups import MultipleEnsembleGroups, SingleEnsembleGroup
 from serena.utilities.ensemble_structures import Sara2SecondaryStructure
 from serena.utilities.ensemble_variation import SourceMFE, EV_Token, EV, EVResult
 from serena.utilities.thread_manager import EV_ThreadProcessor
+
+@dataclass
+class ComparisonLMV():
+    lmv_comp:EV = EV()
+    lmv_mfe:EV = EV()
+    lmv_rel:EV = EV()
 
 class LocalMinimaVariation():
 
     def __init__(self) -> None:
         pass
 
-    def get_local_minima_variation(self, ensemble: MultipleEnsembleGroups, comparison_structure:Sara2SecondaryStructure):        
+    def get_groups_local_minima_variation(self, ensemble: MultipleEnsembleGroups, reference_structure:Sara2SecondaryStructure):        
         #now process all the groups
        
         #source_mfe: SourceMFE = SourceMFE.NONE
@@ -26,12 +33,19 @@ class LocalMinimaVariation():
         
         #single_ensemble_group: List[Sara2StructureList] = [ensemble.group]
         LMV_Thread: EV_ThreadProcessor = EV_ThreadProcessor(stuctures=ensemble.raw_groups,
-                                                              comparison_structure=comparison_structure)
+                                                              comparison_structure=reference_structure)
         result_thread_LMV:EV_Token = LMV_Thread.run_EV()
         group_ev_list: List[EV] = result_thread_LMV.group_results
         group_ev_dict: Dict[int,EV] = result_thread_LMV.group_dict
 
-        result_LMV: EVResult = EVResult(groups_list=ensemble.raw_groups, groups_dict=ensemble.groups_dict, 
-                                              group_values=ensemble.group_values, group_ev_list=group_ev_list, 
-                                              group_ev_dict=group_ev_dict)
-        return result_LMV
+        lmv_results: List[EVResult] = []
+
+        for group_index in range(len(ensemble.groups)):
+            group: SingleEnsembleGroup = ensemble.groups[group_index]
+            ev:EV = group_ev_list[group_index]
+            ev_result:EVResult = EVResult(group_structs=group,
+                                          ev_values=ev)
+            lmv_results.append(ev_result)
+
+        return lmv_results
+
